Initial commit: styling and code cleanup
